refactor: Extract ForgotPasswordForm component for better content streaming
